+++ src/antidote/_constants.py
@@ -1,22 +1,39 @@
-from typing import (Any, Callable, Dict, Generic, Hashable, Optional, TYPE_CHECKING,
-                    Tuple, Type, TypeVar, cast)
+from typing import (Any, Dict, Generic, Hashable, Optional, TYPE_CHECKING, Tuple, Type,
+                    TypeVar, Union, cast, overload)
 
-from ._compatibility.typing import final
+from ._compatibility.typing import final, Protocol
 from ._internal import API
-from ._internal.utils import AbstractMeta, FinalImmutable, FinalMeta, debug_repr
+from ._internal.utils import AbstractMeta, Default, FinalImmutable, FinalMeta, debug_repr
+from ._internal.utils.immutable import Immutable, ImmutableGenericMeta
 from ._providers.lazy import Lazy
 from .core import Container, DependencyDebug, DependencyValue, Scope
 
 T = TypeVar('T')
 
 if TYPE_CHECKING:
-    from mypy_extensions import DefaultNamedArg, DefaultArg
-
-_SENTINEL = object()
+    from .constants import Constants
 
 
+# TODO: Once Python 3.6 support drops, fix this.
+# We're lying to Mypy here. That's not how the actual descriptor, even though it's
+# somewhat close. But inheriting Generic implies not being final anymore in Python 3.6,
+# until PEP 560, and internally there's no need for Generic.
 class Const(Generic[T]):
-    def __get__(self, instance: Any, owner: Any) -> T:  # pragma: no cover
+    __slots__ = ()
+
+    @overload
+    def __get__(self,  # noqa: E704
+                instance: 'Constants',
+                owner: 'Type[Constants]') -> T: ...  # pragma: no cover
+
+    @overload
+    def __get__(self,  # noqa: E704
+                instance: None,
+                owner: 'Type[Constants]') -> 'Const[T]': ...  # pragma: no cover
+
+    def __get__(self,
+                instance: 'Optional[Constants]',
+                owner: 'Type[Constants]') -> object:  # pragma: no cover
         pass
 
 
@@ -26,19 +43,30 @@ class MakeConst(metaclass=FinalMeta):
     def __call__(self,
                  __arg: Optional[object] = None,
                  *,
-                 default: Any = _SENTINEL) -> Const[object]:
+                 default: Any = Default.sentinel) -> Const[object]:
         # Not true yet, but will be changed by ConstantsMeta
         return cast(Const[object], LazyConstToDo(__arg, None, default))
 
-    def __getitem__(self,
-                    tpe: Type[T]
-                    ) -> 'Callable[[DefaultArg(object), DefaultNamedArg(T, "default")], Const[T]]':  # noqa: F821, E501
-        def f(__arg: Optional[object] = None,
-              *,
-              default: T = cast(T, _SENTINEL)) -> Const[T]:
-            return cast(Const[T], LazyConstToDo(__arg, tpe, default))
+    def __getitem__(self, tpe: Type[T]) -> 'MakeTypedConst[T]':
+        return MakeTypedConst(tpe)
 
-        return f
+
+@API.private
+@final
+class MakeTypedConst(Immutable, Generic[T], metaclass=ImmutableGenericMeta):
+    __slots__ = ('__type',)
+    __type: Type[T]
+
+    def __call__(self,
+                 __arg: Optional[object] = None,
+                 *,
+                 default: Union[T, Default] = Default.sentinel) -> Const[T]:
+        if not isinstance(default, (self.__type, Default)):
+            raise TypeError(f"default is not an instance of {self.__type}, "
+                            f"but {type(default)}")
+
+        # Not true yet, but will be changed by ConstantsMeta
+        return cast(Const[T], LazyConstToDo(__arg, self.__type, default))
 
 
 @API.private
@@ -88,26 +116,25 @@ def _configure_constants(cls: ConstantsMeta) -> None:
                     LazyConstDescriptor(
                         name=name,
                         dependency=dependency,
-                        method_name=Constants.get_const.__name__,
+                        method_name=Constants.provide_const.__name__,
                         arg=v.arg,
                         default=v.default,
-                        cast=v.type_ if v.type_ in conf.auto_cast else None))
-
-
-Cast = Callable[[object], object]
+                        type_=v.type_ or object,
+                        auto_cast=v.type_ is not None and v.type_ in conf.auto_cast))
 
 
 @API.private
 @final
 class LazyConstDescriptor(FinalImmutable):
-    __slots__ = ('name', 'dependency', 'method_name', 'arg', 'default', 'cast',
-                 '_cache')
+    __slots__ = ('name', 'dependency', 'method_name', 'arg', 'default', 'type_',
+                 'auto_cast', '_cache')
     name: str
     dependency: Hashable
     method_name: str
     arg: object
     default: object
-    cast: Cast
+    type_: type
+    auto_cast: bool
     _cache: str
 
     def __init__(self,
@@ -117,14 +144,18 @@ class LazyConstDescriptor(FinalImmutable):
                  method_name: str,
                  arg: object,
                  default: object,
-                 cast: Cast = None):
+                 type_: type,
+                 auto_cast: bool
+                 ):
+        assert isinstance(default, (Default, type_))
         super().__init__(
             name=name,
             dependency=dependency,
             method_name=method_name,
             arg=arg,
             default=default,
-            cast=cast or (lambda x: x),
+            type_=type_,
+            auto_cast=auto_cast,
             _cache=f"__antidote_dependency_{hex(id(self))}"
         )
 
@@ -133,19 +164,25 @@ class LazyConstDescriptor(FinalImmutable):
             try:
                 return getattr(owner, self._cache)
             except AttributeError:
-                # TODO: Waiting for a fix: https://github.com/python/mypy/issues/6910
-                _cast = cast(Cast, getattr(self, 'cast'))
                 dependency = LazyConst(self)
                 setattr(owner, self._cache, dependency)
                 return dependency
-        _cast = cast(Cast, getattr(self, 'cast'))
         try:
-            return _cast(getattr(instance, self.method_name)(name=self.name,
-                                                             arg=self.arg))
+            value = getattr(instance, self.method_name)(name=self.name,
+                                                        arg=self.arg)
         except LookupError:
-            if self.default is not _SENTINEL:
+            if self.default is not Default.sentinel:
                 return self.default
             raise
+
+        if self.auto_cast:
+            value = self.type_(value)
+
+        if not isinstance(value, self.type_):
+            raise TypeError(f"Constant {self.name} is not an instance of {self.type_}, "
+                            f"but {type(value)}")
+
+        return value
 
 
 @API.private