+++ tests/core/test_utils.py
@@ -1,3 +1,5 @@
+import pytest
+
 from antidote import world
 from antidote.core import LazyDependency, DependencyDebug, DependencyValue, Scope
 
@@ -5,9 +7,21 @@ from antidote.core import LazyDependency, DependencyDebug, DependencyValue, Scop
 def test_dependency():
     with world.test.empty():
         world.test.singleton('x', object())
-        d = LazyDependency('x')
+        d = LazyDependency('x', object)
         assert d.unwrapped == 'x'
         assert d.get() is world.get('x')
+
+    class A:
+        pass
+
+    with world.test.empty():
+        world.test.singleton('a', A())
+        world.test.singleton('x', object())
+
+        assert LazyDependency('a', A).get() is world.get('a')
+
+        with pytest.raises(TypeError):
+            LazyDependency('x', A).get()
 
 
 def test_dependency_instance():
@@ -42,15 +56,3 @@ def test_dependency_debug():
                                   scope=Scope.singleton(),
                                   wired=[1],
                                   dependencies=[20])
-
-
-def test_dependency_hash_eq():
-    d1 = LazyDependency('same')
-    d2 = LazyDependency('same')
-    x = LazyDependency('different')
-
-    assert hash(d1) == hash(d2)
-    assert d1 == d2
-
-    assert hash(d1) != hash(x)
-    assert d1 != x