+++ tests/mypy_typing/test_readme.py
@@ -11,21 +11,21 @@ def test_readme_simple():
     # from typing_extensions import Annotated # Python < 3.9
 
     class Conf(Constants):
-        DB_HOST = const[str]('localhost:6789')
-        DB_HOST_WITHOUT_TYPE_HINT = const('localhost:6789')
+        DB_HOST = const[str]('localhost:5432')
+        DB_HOST_WITHOUT_TYPE_HINT = const('localhost:5432')
 
-    class Database(Service):  # Defined as a Service, so injectable.
+    class Database(Service):  # Declared as a Service
         def __init__(self, host: Annotated[str, Get(Conf.DB_HOST)]):
-            self._host = host  # <=> Conf().get('host')
+            self._host = host  # <=> Conf().DB_HOST <=> 'localhost:5432'
 
-    @inject  # Nothing is injected implicitly.
+    @inject
     def f(db: Provide[Database] = None):
         # Defaulting to None allows for MyPy compatibility but isn't required to work.
         assert db is not None
         pass
 
-    f()  # works !
-    f(Database('localhost:6789'))  # but you can still use the function normally
+    f()  # yeah !
+    f(Database('localhost:5432'))  # override injections
 
     # You can also retrieve dependencies by hand
     world.get(Conf.DB_HOST)
@@ -36,7 +36,7 @@ def test_readme_simple():
     ######################
 
     class Database(Service):
-        @inject(dependencies={'host': Conf.DB_HOST})
+        @inject({'host': Conf.DB_HOST})
         def __init__(self, host: str):
             self._host = host
 
@@ -45,8 +45,7 @@ def test_readme_simple():
         assert db is not None
         pass
 
-    # Or with auto_provide=True, all class type hints will be treated as dependencies.
-    # you can also explicitly say which classes with `auto_provide=[Database]`.
+    # auto_provide => Class type hints are treated as dependencies.
     @inject(auto_provide=True)
     def f(db: Database = None):
         assert db is not None
@@ -54,31 +53,31 @@ def test_readme_simple():
 
 
 def test_readme():
-    """
-    Simple example where a MovieDB interface is defined which can be used
-    to retrieve the best movies. In our case the implementation uses IMDB
-    to dot it.
-    """
+
+    # Some library.py
+    class ImdbAPI:
+        """ Class from an external library. """
+
+        def __init__(self, host: str, port: int, api_key: str):
+            pass
+
+    ######################
+
+    class MovieDB:
+        """ Interface """
+
+    ######################
+
     from antidote import (Constants, factory, inject, world, const, Service,
                           implementation, Get, From)
     # from typing import Annotated
     # from typing_extensions import Annotated # Python < 3.9
 
-    class MovieDB:
-        """ Interface """
-
-    class ImdbAPI:
-        """ Class from an external library. """
-
-        def __init__(self, *args, **kwargs):
-            pass
-
     class Conf(Constants):
+        # with str/int/float, the type hint is enforced. Can be removed or extend to
+        # support Enums.
         IMDB_HOST = const[str]('imdb.host')
-        # Constants will by default automatically enforce the cast to int,
-        # float and str. Can be removed or extended to support Enums.
         IMDB_PORT = const[int]('imdb.port')
-        # But specifying a type is not required at all, it's mostly to help Mypy.
         IMDB_API_KEY = const('imdb.api_key')
 
         def __init__(self):
@@ -94,18 +93,16 @@ def test_readme():
                 }
             }
 
-        def get_const(self, name: str, arg: str):
-            from functools import reduce
-            # self.get('a.b') <=> self._raw_conf['a']['b']
-            return reduce(dict.get, arg.split('.'), self._raw_conf)  # type: ignore
+        def provide_const(self, name: str, arg: str):
+            root, key = arg.split('.')
+            return self._raw_conf[root][key]
 
-    # Provides ImdbAPI, as defined by the return type annotation.
+    # Provides ImdbAPI as defined by the return type annotation.
     @factory
     def imdb_factory(host: Annotated[str, Get(Conf.IMDB_HOST)],
                      port: Annotated[int, Get(Conf.IMDB_PORT)],
                      api_key: Annotated[str, Get(Conf.IMDB_API_KEY)]
                      ) -> ImdbAPI:
-        # Here host = Conf().get('imdb.host')
         return ImdbAPI(host=host, port=port, api_key=api_key)
 
     @implementation(MovieDB)
@@ -113,11 +110,12 @@ def test_readme():
         return IMDBMovieDB  # dependency to be provided for MovieDB
 
     class IMDBMovieDB(MovieDB, Service):
-        # New instance each time
-        __antidote__ = Service.Conf(singleton=False)
+        __antidote__ = Service.Conf(singleton=False)  # New instance each time
 
         def __init__(self, imdb_api: Annotated[ImdbAPI, From(imdb_factory)]):
             self._imdb_api = imdb_api
+
+    ######################
 
     @inject
     def f(movie_db: Annotated[MovieDB, From(current_movie_db)] = None):
@@ -131,7 +129,7 @@ def test_readme():
     @factory
     @inject([Conf.IMDB_HOST, Conf.IMDB_PORT, Conf.IMDB_API_KEY])
     def imdb_factory(host: str, port: int, api_key: str) -> ImdbAPI:
-        return ImdbAPI(host=host, port=port, api_key=api_key)
+        return ImdbAPI(host, port, api_key)
 
     class IMDBMovieDB(MovieDB, Service):
         __antidote__ = Service.Conf(singleton=False)
@@ -149,16 +147,15 @@ def test_readme():
 
     conf = Conf()
     f(IMDBMovieDB(imdb_factory(
-        # The class attributes will retrieve the actual value when called on a instance.
-        # Hence this is equivalent to conf.get('imdb.host'), making your tests easier.
+        # constants can be retrieved directly on an instance
         host=conf.IMDB_HOST,
         port=conf.IMDB_PORT,
-        api_key=conf.IMDB_API_KEY,  # <=> conf.get('imdb.api_key')
+        api_key=conf.IMDB_API_KEY,
     )))
 
     ######################
 
-    # When testing you can also override locally some dependencies:
+    # Override locally some dependencies:
     with world.test.clone(keep_singletons=True):
         world.test.override.singleton(Conf.IMDB_HOST, 'other host')
         f()